+++ src/command_getid.hpp
@@ -23,9 +23,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 */
 
+#include <map>
+#include <set>
 #include <string>
 #include <vector>
 
+#include <osmium/fwd.hpp>
 #include <osmium/osm/entity_bits.hpp>
 #include <osmium/osm/item_type.hpp>
 #include <osmium/osm/types.hpp>
@@ -34,28 +37,42 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 class CommandGetId : public Command, public with_single_osm_input, public with_osm_output {
 
-    std::vector<osmium::object_id_type> m_ids[3];
-
     osmium::item_type m_default_item_type = osmium::item_type::node;
+
+    bool m_add_referenced_objects = false;
+    bool m_work_with_history = false;
+
+    std::set<osmium::object_id_type> m_ids[3];
+
+    std::set<osmium::object_id_type>& ids(osmium::item_type type) noexcept;
+    const std::set<osmium::object_id_type>& ids(osmium::item_type type) const noexcept;
+
+    void parse_and_add_id(const std::string& s);
+
+    void read_id_osm_file(const std::string& file_name);
+    void read_id_file(std::istream& stream);
+
+    osmium::osm_entity_bits::type get_needed_types() const;
+    bool no_ids() const;
+    size_t count_ids() const;
+
+    void find_referenced_objects();
+
+    void add_nodes(const osmium::Way& way);
+    void add_members(const osmium::Relation& relation);
+
+    void mark_rel_ids(const std::multimap<osmium::object_id_type, osmium::object_id_type>& rel_in_rel, osmium::object_id_type id);
+    bool find_relations_in_relations();
+    void find_nodes_and_ways_in_relations();
+    void find_nodes_in_ways();
 
 public:
 
     CommandGetId() = default;
 
-    std::vector<osmium::object_id_type>& ids(osmium::item_type type) noexcept;
-    const std::vector<osmium::object_id_type>& ids(osmium::item_type type) const noexcept;
-
-    void sort_unique(osmium::item_type type);
-
-    void parse_and_add_id(const std::string& s);
-
-    void read_id_file(std::istream& stream);
-
     bool setup(const std::vector<std::string>& arguments) override final;
 
     void show_arguments() override final;
-
-    osmium::osm_entity_bits::type get_needed_types() const;
 
     bool run() override final;